+++ lapce-core/src/word.rs
@@ -407,6 +407,43 @@ impl<'a> WordCursor<'a> {
         let start = self.prev_code_boundary();
         (start, end)
     }
+
+    /// Return the enclosing brackets of the current position
+    ///
+    /// **Example**:
+    ///
+    ///```rust
+    /// # use lapce_core::word::WordCursor;
+    /// # use xi_rope::Rope;
+    /// let text = "outer {{inner} world";
+    /// let rope = Rope::from(text);
+    /// let mut cursor = WordCursor::new(&rope, 10);
+    /// let (start, end) = cursor.find_enclosing_pair().unwrap();
+    /// assert_eq!(start, 7);
+    /// assert_eq!(end, 13)
+    ///```
+    pub fn find_enclosing_pair(&mut self) -> Option<(usize, usize)> {
+        let old_offset = self.inner.pos();
+        while let Some(c) = self.inner.next_codepoint() {
+            if matching_pair_direction(c) == Some(false) {
+                let closing_bracket_offset = self.inner.pos() - 1;
+                self.inner.set(closing_bracket_offset);
+                if let Some(opening_bracket_offset) = self.match_pairs() {
+                    if (opening_bracket_offset..closing_bracket_offset)
+                        .contains(&old_offset)
+                    {
+                        return Some((
+                            opening_bracket_offset,
+                            closing_bracket_offset,
+                        ));
+                    } else {
+                        self.inner.set(closing_bracket_offset + 1);
+                    }
+                }
+            }
+        }
+        None
+    }
 }
 
 /// Return the [`CharClassification`] of the input character
@@ -638,5 +675,32 @@ mod test {
 
         assert_eq!(position, Some(7));
         assert_eq!(&text[..position.unwrap()], "violet ");
+    }
+
+    #[test]
+    fn find_pair_should_return_positions() {
+        let text = "violet (are) blue";
+        let rope = Rope::from(text);
+        let mut cursor = WordCursor::new(&rope, 9);
+        let positions = cursor.find_enclosing_pair();
+        assert_eq!(positions, Some((7, 11)));
+    }
+
+    #[test]
+    fn find_pair_should_return_next_pair() {
+        let text = "violets {are (blue)}";
+        let rope = Rope::from(text);
+        let mut cursor = WordCursor::new(&rope, 11);
+        let positions = cursor.find_enclosing_pair();
+        assert_eq!(positions, Some((8, 19)))
+    }
+
+    #[test]
+    fn find_pair_should_return_none() {
+        let text = "violet (are) blue";
+        let rope = Rope::from(text);
+        let mut cursor = WordCursor::new(&rope, 1);
+        let positions = cursor.find_enclosing_pair();
+        assert_eq!(positions, None);
     }
 }