+++ lapce-data/src/document.rs
@@ -24,7 +24,8 @@ use lapce_core::{
     register::{Clipboard, Register, RegisterData},
     selection::{SelRegion, Selection},
     style::line_styles,
-    syntax::{edit::SyntaxEdit, Syntax},
+    syntax::edit::SyntaxEdit,
+    syntax::{util::matching_pair_direction, Syntax},
     word::WordCursor,
 };
 use lapce_rpc::{
@@ -2780,5 +2781,31 @@ impl Document {
         } else {
             None
         }
+    }
+
+    pub fn find_enclosing_brackets(&self, offset: usize) -> Option<(usize, usize)> {
+        let char_at_cursor = match self.buffer().char_at_offset(offset) {
+            Some(c) => c,
+            None => return None,
+        };
+
+        if let Some(syntax) = self.syntax() {
+            if matching_pair_direction(char_at_cursor).is_some() {
+                if let Some(new_offset) = syntax.find_matching_pair(offset) {
+                    return Some((offset, new_offset));
+                }
+            } else {
+                return syntax.find_enclosing_pair(offset);
+            }
+        } else {
+            let mut cursor = WordCursor::new(self.buffer.text(), offset);
+            if matching_pair_direction(char_at_cursor).is_some() {
+                let new_offset = cursor.match_pairs().unwrap_or(offset);
+                return Some((offset, new_offset));
+            } else {
+                return cursor.find_enclosing_pair();
+            }
+        }
+        None
     }
 }